GPS: using 16-bit integers for course over ground

+++ openrtx/include/core/gps.h
@@ -52,8 +52,8 @@ typedef struct
     int32_t    longitude;            // Longitude coordinates
     int16_t    altitude;             // Antenna altitude above mean sea level (geoid) in m
     uint16_t   speed;                // Ground speed in km/h
-    float      tmg_mag;              // Course over ground, degrees, magnetic
-    float      tmg_true;             // Course over ground, degrees, true
+    int16_t    tmg_mag;              // Course over ground, degrees, magnetic
+    int16_t    tmg_true;             // Course over ground, degrees, true
 }
 gps_t;
